fix(module): module bootstrap flow changed to initialize module as soon as possible
