--- controller.go
@@ -1,6 +1,0 @@
-package flow
-
-// ControllerRouter interface allows controllers to define their routing logic
-type ControllerRouter interface {
-	Routes(*Router)
-}
